requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
djangorestframework-jwt==1.11.0